==> 4 arxiv.ipynb <==
import pandas as pd
df = pd.read_csv("papers_clean.tsv", sep="\t", encoding="utf-8")
df.head(2)

import requests
from IPython.core.display import display, HTML

url = 'https://arxiv.org/abs/2506.17298'
response = requests.get(url)
if response.status_code == 200:
    print("Successfully fetched the page.")
    
with open("arxiv_page.html", "w", encoding="utf-8") as file:
    file.write(response.text)

import re

input_file = "arxiv_page.html"

with open(input_file, "r", encoding="utf-8") as f:
    html = f.read()

pattern = re.compile(
    r'(<div class="leftcolumn">.*?<div style="clear:both;"></div>)',
    re.DOTALL
)

match = pattern.search(html)
if match:
    extracted = match.group(1)

print(extracted)

from arxiv import arxiv_to_json

arxiv_id, data = arxiv_to_json(extracted)

data
